BUG: Fixup the quantile promotion fixup
Sorry, I incorrectly reverted a line and made things maybe worse :(.
Basically it now *always* returned float64 (at least for most methods).

I am surprised that we had not tests for this at all.

I suppose this needs backporting, which isn't great, but...

+++ numpy/lib/_function_base_impl.py
@@ -4576,7 +4576,7 @@ def _get_gamma(virtual_indexes, previous_indexes, method):
     gamma = method["fix_gamma"](gamma, virtual_indexes)
     # Ensure both that we have an array, and that we keep the dtype
     # (which may have been matched to the input array).
-    return np.asanyarray(gamma)
+    return np.asanyarray(gamma, dtype=virtual_indexes.dtype)
 
 
 def _lerp(a, b, t, out=None):
